requirements.txt
allennlp==1.0.0rc1
allennlp-models==1.0.0rc1
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
blis==0.4.1
boto3==1.12.46
botocore==1.15.46
CacheControl==0.12.6
CacheMan==2.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
conllu==2.3.2
contextlib2==0.6.0
CyHunspell==1.3.4
cymem==2.0.3
decorator==4.4.2
dill==0.2.9
distlib==0.3.0
distro==1.4.0
Django==3.0.7
django-cors-headers==3.3.0
docutils==0.15.2
en-core-web-sm==2.2.5
filelock==3.0.12
flaky==3.6.1
ftfy==4.4.3
future==0.18.2
h5py==2.10.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
ipaddr==2.2.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
jmespath==0.9.5
joblib==0.14.1
jsonnet==0.15.0
jsonpickle==1.4.1
lockfile==0.12.2
more-itertools==8.2.0
msgpack==0.6.2
murmurhash==1.0.2
nltk==3.5
numpy==1.18.3
overrides==2.8.0
packaging==20.3
parso==0.7.0
pathlib==1.0.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
progress==1.5
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
py-rouge==1.1
Pygments==2.6.1
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
regex==2017.11.9
requests==2.23.0
responses==0.10.14
retrying==1.3.3
s3transfer==0.3.3
sacremoses==0.0.41
scikit-learn==0.22.2.post1
scipy==1.4.1
semantic-version==2.8.4
sentencepiece==0.1.86
six==1.14.0
spacy==2.2.4
sqlparse==0.3.1
srsly==1.0.2
tensorboardX==2.0
thinc==7.4.0
tokenizers==0.5.2
toolz==0.10.0
torch==1.4.0
tqdm==4.45.0
traitlets==4.3.3
transformers==2.8.0
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
word2number==1.1
zipp==3.1.0
